Refactor typemap and store on interpreter
Typemaps are no longer stored in a map with type_index keys. When a
usertype class is registered, a unique id is generated for the type
using a static incrementing counter. This id is associated with the
type name and stored in a hash on the perl interpreter. These ids
should be unique for the execution of the whole program.

Note these are stored in a hash with string keys. An array with
numerical type id indexes cannot be used because it's not guaranteed
that types would be registered for index lookup.

This finalizes making perlbind stateless so its interpreter object
doesn't need to persist.