Use SV for some croak error messages
Address sanitizers detected memory leaks from the std::string objects
on the stack before calls to Perl_croak. It might be a false positive
if the stack properly unwinds to destroy them, but croaking with the
error message held in a mortal SV may be safer.

Other places that use Perl_croak either don't have any objects on the
stack before the croak or temporarily construct for the format string.

There is an array object in the same scope that appears to be properly
destructed after croaking.