Mark callbacks extern C
This fixes a stack corruption bug in MSVC Release builds

If the static package::xsub function croaked, objects used inside the
overload loops would get destructed again during croak unwinding. This
was was causing the last used iterator object to decrement twice which
corrupted the last element in the overloads array.

This only occured when a function's virtual method was called in the
loop. This needs more investigation in case there's some undefined
behavior occuring. It might also have something to do with the callback
being called from the perl dll.