Use exceptions to prevent xsub leaks
The extern "C" stack corruption fix exposed a serious issue with the
use of croaks that could lead to memory leaks.

The xsub callback must croak and not throw exceptions as it's called by
a C library. Since croak uses longjmp to return back to its eval(), the
C++ stack does not unwind and objects will not be destroyed. This caused
the array object in the xsub callback to leak references after a croak.
This was the issue detected by address sanitizer for the std::string
error message leak.

This replaces all Perl_croak calls with thrown exceptions when a C++
error occurs. The xsub callback will now catch exceptions and croak
to make sure the C++ stack is properly unwound before returning to perl.